Use a struct to hold config
Why:
* This lets us indicate taht query and filename are some of the same
configuration. We can now reference each of them as the same
configuration

This change is addresses this by:
* using clone() to copy over the args (note: this has some runtime
tradeoffs in that we're now copying stings around in memory. that seems
like not such a bad thing because our values will be small but oculd be
a bad thing given huge input sizes)